retirer les liens pour les non admins + recuperation de mots de passe + envot de mails

+++ src/app/pages/welcome/autorisations/autorisations.component.ts
@@ -3,6 +3,8 @@ import { Autorisation } from '../../../Models/Autorisations';
 import { HttpClient } from '@angular/common/http';
 import { ApiConceptsEtTravauxService } from '../../../Services/api-concepts-et-travaux.service';
 import { NzButtonSize } from 'ng-zorro-antd/button';
+import { AuthService } from '../../../Services/auth.service';
+import { NzMessageService } from 'ng-zorro-antd/message';
 
 @Component({
   selector: 'app-autorisations',
@@ -38,7 +40,7 @@ export class AutorisationsComponent {
     }
   ];
 
-  constructor(private http: HttpClient,private userService: ApiConceptsEtTravauxService) { }
+  constructor(private http: HttpClient,private authService: AuthService,private message: NzMessageService,private userService: ApiConceptsEtTravauxService) { }
 
   ngOnInit(): void {
     this.loadAutorisations();
@@ -58,15 +60,23 @@ export class AutorisationsComponent {
   }
 
   deleteAutorisation(autoId: number) {
-    this.userService.deleteAutorisation(autoId).subscribe(
-      () => {
-        console.log('Utilisateur supprimé avec succès');
-        // Mettez ici le code pour actualiser la liste des Autorisations si nécessaire
-        this.loadAutorisations();
-      },
-      (error) => {
-        console.error('Erreur lors de la suppression de l\'utilisateur :', error);
-      }
-    );
+    if (this.authService.isAdmin()) {
+      this.userService.deleteAutorisation(autoId).subscribe(
+        () => {
+          console.log('Utilisateur supprimé avec succès');
+          // Mettez ici le code pour actualiser la liste des Autorisations si nécessaire
+          this.loadAutorisations();
+        },
+        (error) => {
+          console.error('Erreur lors de la suppression de l\'utilisateur :', error);
+        }
+      );
+      return true
+    } else {
+      this.message.info( `Vous n'avez pas assez de privilèges pour acceder à cette page et/ou ce n'est pas votre compte`);
+      return false;
+    }
+
+    
   }
 }
